Update __init__.py
Replace the deprecated `async_get_registry` to the new `async_get` method! You cannot await `async_get` so it was removed. If there's a valid reason it should wait on a promise, let me know and I'm we can figure something out. Let me know if this closes this long-standing issue for you 👍🏼